allow set sampling rate and sampling mode, update updateContextSourceList

+++ app/src/main/java/edu/umich/si/inteco/minuku/Constants.java
@@ -66,13 +66,13 @@ public class Constants {
 
 
 	/**constant**/
-	public static final int MILLISECONDS_PER_SECOND = 1000;
+	public static final long MILLISECONDS_PER_SECOND = 1000;
 	public static final int SECONDS_PER_MINUTE = 60;
 	public static final int MINUTES_PER_HOUR = 60;
 	public static final int HOURS_PER_DAY = 24;
-	public static final int MILLISECONDS_PER_DAY = HOURS_PER_DAY *MINUTES_PER_HOUR*SECONDS_PER_MINUTE*MILLISECONDS_PER_SECOND;
-    public static final int MILLISECONDS_PER_HOUR = MINUTES_PER_HOUR*SECONDS_PER_MINUTE*MILLISECONDS_PER_SECOND;
-    public static final int MILLISECONDS_PER_MINUTE = SECONDS_PER_MINUTE*MILLISECONDS_PER_SECOND;
+	public static final long MILLISECONDS_PER_DAY = HOURS_PER_DAY *MINUTES_PER_HOUR*SECONDS_PER_MINUTE*MILLISECONDS_PER_SECOND;
+    public static final long MILLISECONDS_PER_HOUR = MINUTES_PER_HOUR*SECONDS_PER_MINUTE*MILLISECONDS_PER_SECOND;
+    public static final long MILLISECONDS_PER_MINUTE = SECONDS_PER_MINUTE*MILLISECONDS_PER_SECOND;
     public static final String DATE_FORMAT_NOW = "yyyy-MM-dd HH:mm:ss Z";
     public static final String DATE_FORMAT_NOW_NO_ZONE = "yyyy-MM-dd HH:mm:ss";
 	public static final String DATE_FORMAT_NOW_DAY = "yyyy-MM-dd";
